fix(build): clang -Werror fixes for the Linux/macOS release legs
The windows-x64/x86 legs build under MSVC, which is lax about several
things clang -Werror rejects, so the ubuntu (ninja-clang) and macOS legs
have been red since before rc29. None of these are reachable on MSVC:

- ace_exports.cpp: <cmath> wasn't included, so std::floor was undeclared
  under libc++ (MSVC pulls it in transitively).
- The 4-byte LE decode `b[0]|(b[1]<<8)|...` promotes to int and assigns
  to uint32_t -> -Wsign-conversion. Cast each byte to uint32_t.
- build_system_dbf / dispatch_sp_builtin sit inside an extern "C" block
  and need C++ linkage to return std::string; the `extern "C++" static`
  spelling that MSVC accepts is rejected by clang ("cannot combine with
  previous extern"). Drop `static`, keep `extern "C++"`.
- data_dict.cpp: uint16_t slen = a | (b<<8) narrows int->uint16_t under
  -Wimplicit-int-conversion. Wrap in an explicit static_cast<uint16_t>.

+++ src/engine/data_dict.cpp
@@ -390,9 +390,9 @@ util::Result<void> DataDict::load_add_binary_(const std::string& buf) {
                     // holds the continuation.  Read as many bytes as fit.
                     auto read_lstr = [&](std::string& out) {
                         if (pos + 2 > PL) return;
-                        uint16_t slen =
+                        uint16_t slen = static_cast<uint16_t>(
                             static_cast<uint8_t>(buf[PS + pos]) |
-                            (static_cast<uint16_t>(static_cast<uint8_t>(buf[PS + pos + 1])) << 8);
+                            (static_cast<uint16_t>(static_cast<uint8_t>(buf[PS + pos + 1])) << 8));
                         pos += 2;
                         if (slen == 0 || slen == 0xFFFF) return;
                         std::size_t readable = std::min<std::size_t>(slen, PL - pos);
